# Kernel+LibC+LibCore: Optimize SpaceAnalyzer startup time