--- a/Userland/Libraries/LibC/fcntl.h
+++ b/Userland/Libraries/LibC/fcntl.h
@@ -40,6 +40,7 @@ __BEGIN_DECLS
 int creat(const char* path, mode_t);
 int open(const char* path, int options, ...);
 #define AT_FDCWD -100
+#define AT_SYMLINK_NOFOLLOW 0x100
 int openat(int dirfd, const char* path, int options, ...);
 
 int fcntl(int fd, int cmd, ...);