--- a/Userland/Libraries/LibC/stat.cpp
+++ b/Userland/Libraries/LibC/stat.cpp
@@ -6,6 +6,7 @@
 
 #include <assert.h>
 #include <errno.h>
+#include <fcntl.h>
 #include <stdio.h>
 #include <string.h>
 #include <sys/stat.h>
@@ -50,30 +51,35 @@ int mkfifo(const char* pathname, mode_t mode)
     return mknod(pathname, mode | S_IFIFO, 0);
 }
 
-static int do_stat(const char* path, struct stat* statbuf, bool follow_symlinks)
+static int do_stat(int dirfd, const char* path, struct stat* statbuf, bool follow_symlinks)
 {
     if (!path) {
         errno = EFAULT;
         return -1;
     }
-    Syscall::SC_stat_params params { { path, strlen(path) }, statbuf, follow_symlinks };
+    Syscall::SC_stat_params params { dirfd, { path, strlen(path) }, statbuf, follow_symlinks };
     int rc = syscall(SC_stat, &params);
     __RETURN_WITH_ERRNO(rc, rc, -1);
 }
 
 int lstat(const char* path, struct stat* statbuf)
 {
-    return do_stat(path, statbuf, false);
+    return do_stat(AT_FDCWD, path, statbuf, false);
 }
 
 int stat(const char* path, struct stat* statbuf)
 {
-    return do_stat(path, statbuf, true);
+    return do_stat(AT_FDCWD, path, statbuf, true);
 }
 
 int fstat(int fd, struct stat* statbuf)
 {
     int rc = syscall(SC_fstat, fd, statbuf);
     __RETURN_WITH_ERRNO(rc, rc, -1);
+}
+
+int fstatat(int fd, const char* path, struct stat* statbuf, int flags)
+{
+    return do_stat(fd, path, statbuf, !(flags & AT_SYMLINK_NOFOLLOW));
 }
 }