--- a/Userland/Libraries/LibC/sys/stat.h
+++ b/Userland/Libraries/LibC/sys/stat.h
@@ -77,6 +77,7 @@ int mkfifo(const char* pathname, mode_t);
 int fstat(int fd, struct stat* statbuf);
 int lstat(const char* path, struct stat* statbuf);
 int stat(const char* path, struct stat* statbuf);
+int fstatat(int fd, const char* path, struct stat* statbuf, int flags);
 
 inline dev_t makedev(unsigned int major, unsigned int minor) { return (minor & 0xffu) | (major << 8u) | ((minor & ~0xffu) << 12u); }
 inline unsigned int major(dev_t dev) { return (dev & 0xfff00u) >> 8u; }