By implementing fstatat and using it in SpaceAnalyzer, its startup time went from:

- cold cache: ~1331ms
- hot cache: ~636ms

to

- cold cache: ~812ms
- hot cache ~324ms.